backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
Flask-Migrate
Flask-cors
Flask-JWT-Extended